requirements.txt
streamlit == 1.24.1
google-generativeai == 0.3.0